MinimumSimilarity should be a double #411

+++ src/Nest/DSL/Query/FuzzyLikeThisDescriptor.cs
@@ -21,7 +21,7 @@ namespace Nest
 		[JsonProperty(PropertyName = "max_query_terms")]
 		internal int? _MaxQueryTerms { get; set; }
 		[JsonProperty(PropertyName = "min_similarity")]
-		internal int? _MinSimilarity { get; set; }
+		internal double? _MinSimilarity { get; set; }
 		[JsonProperty(PropertyName = "prefix_length")]
 		internal int? _PrefixLength { get; set; }
 		[JsonProperty(PropertyName = "boost")]
@@ -59,7 +59,7 @@ namespace Nest
 			this._IgnoreTermFrequency = ignore;
 			return this;
 		}
-		public FuzzyLikeThisDescriptor<T> MinimumSimilarity(int minSimilarity)
+		public FuzzyLikeThisDescriptor<T> MinimumSimilarity(double minSimilarity)
 		{
 			this._MinSimilarity = minSimilarity;
 			return this;
